If mx_epics_pv_connect() times out or otherwise fails, the code now does this
    (void) ca_clear_channel( new_channel_id );
    pv->channel_id = NULL;

The call to ca_clear_channel() prevents us from having memory leaks due to
abandoned channel id structures.  Setting pv->channel_id to NULL allows us
to start over fresh with the next call to mx_epics_pv_connect() for this PV.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@2287 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_io.c
@@ -935,7 +935,8 @@ mx_file_has_changed( MX_FILE_MONITOR *monitor )
 		(void) mx_error( MXE_FILE_IO_ERROR, fname,
 		"Short read of %lu bytes seen when %lu bytes were expected "
 		"from inotify fd %d for file '%s'",
-			read_length, num_bytes_available,
+			(unsigned long) read_length,
+			(unsigned long) num_bytes_available,
 			linux_monitor->inotify_file_descriptor,
 			monitor->filename );
 	}
